functions/azure-cosmosdb-changefeed-py/PyFunctionApp/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

arrow
azure-functions